font fixes, supervisord fixes

+++ cookbooks/webscreenshots/recipes/debian.rb
@@ -3,7 +3,7 @@
 end
 
 # for PIL
-["libjpeg-dev",].each do |dev_pkg|
+["libjpeg-dev"].each do |dev_pkg|
   package dev_pkg
 end
 
@@ -14,20 +14,14 @@ bash "run ttf-mscorefonts-installer" do
   echo ttf-mscorefonts-installer msttcorefonts/accepted-mscorefonts-eula select true | debconf-set-selections
   apt-get install -y ttf-mscorefonts-installer
   EOS
-#  not_if "test -e /opt/phantomjs-#{node["phantomjs"]["version"]}-linux-x86_64"
+  not_if "test -e /opt/phantomjs-#{node["phantomjs"]["version"]}-linux-#{node["webscreenshots"]["phantomjs"]["arch"]}"
 end
 
-## install mac fonts
-## http://grasshopperpebbles.com/ubuntu/ubuntu-convert-mac-dfont-files-into-ttf-using-fondu/
-#node["phantomjs"]["macfonts"].each do |ttf_file|
-#  cookbook_file "/usr/share/fonts/truetype/#{ttf_file}" do
-#    source ttf_file
-#    owner "root"
-#    group "root"
-#    mode "0644"
-#    end
-#end
-#
-#bash "clear fontchache" do
-#  code "fc-cache -fv"
-#end
+cookbook_file "/etc/fonts/local.conf" do
+  source "localfonts.conf"
+  mode 0644
+end
+
+bash "clear fontchache" do
+  code "fc-cache -fv"
+end
